Extract guides into their own component in a folder, add template, update README
