+++ publisher/src/components/HelpCenter/README.md
@@ -1,18 +1,18 @@
-## How to add or update Help Center Guides
+## How to add, update or delete Help Center Guides
+
 
 ### If you'd like to UPDATE a Help Center guide:
-  1. Go to the `HelpCenterGuides.tsx` file
-  2. Locate & update the guide (functional component) directly
+  1. Go to the `Guides` folder in this directory
+  2. Locate the guide file you'd like to update
+  2. Update the body of the component directly (using HTML/JSX/styled-components)
 
 
-  ###### HelpCenterGuides.tsx
+  ###### ./Guides/Guide.tsx (for example)
 
 ```jsx
-... // Other guides
-
 export const Guide = () => (
-  // Make edits within here
   <> 
+    {/* Make edits within here */}
     <Styled.SectionWrapper>
       <Styled.SectionParagraph>
         ...
@@ -29,26 +29,29 @@ export const Guide = () => (
 
 ### If you'd like to ADD a Help Center guide:
 
-  1. Go to the `HelpCenterGuides.tsx` file
-  2. Add a functional component guide that renders the desired JSX
+  1. Go to the `Guides` folder in this directory
+  2. Duplicate the `TEMPLATE.tsx` file
+  2. Update the file name and component name (make sure they match) and update the body of the component with the desired JSX
   3. Then, go to the `HelpCenterSetup.tsx` file
   4. Import your newly created component
   5. Update the `helpCenterGuideStructure` object and create a guide object for your newly created guide under the `guides` property for either the Publisher/Dashboard app with the following properties:
         * **key**: a unique String that distinguishes this guide
         * **category**: the section that this guide falls under in the Publisher directory page (use the `PublisherGuideCategories` enum)
-        * **label**: the display name
-        * **caption**: the description of the guide
+        * **label**: the display name (will be used to render the title)
+        * **caption**: the description of the guide (will be used to render the caption)
         * **path**: URL pathname for this guide
         * **element**: the guide component you created
         * **relevantGuides**: a list of relevant guide keys
           * **Important note to ensure proper linking**: if a relevant guide belongs to a different app guide, please prepend `<publisher or dashboard>/` to the key. For example, if the Dashboards guide (within the Dashboard app guide) links to the Explore Data guide (within the Publisher app guide) the relevant guide key would be `publisher/explore-data`. If it links to a guide within the same app guide (e.g. Agency Settings guide links to Explore Data guide - both Publisher guides), you can just add the guide key (e.g. `explore-data`) without prepending anything.
 
-###### HelpCenterGuides.tsx
+
+**NOTE**: The title, caption and relevant links will be automatically rendered based on the object created in `HelpCenterSetup.tsx`, so all you need to worry about in a guide component is the body of the guide.
+
+###### ./Guides/TEMPLATE.tsx > ./Guides/NewGuide.tsx
 
 ```jsx
-... // Other guides
+...
 
-// New Guide Template
 export const NewGuide = () => (
   <>
     <Styled.SectionWrapper>
@@ -72,13 +75,9 @@ export const NewGuide = () => (
 
 ###### HelpCenterSetup.tsx
 ```jsx
-import {
-  AccountSetupGuide,
-  ExploreDataGuide,
-  HelpCenterPublisher,
-  NewGuide,
-  ...
-} from ".";
+...
+
+import { NewGuide } from "./Guides/NewGuide";
 
 export const helpCenterGuideStructure: HelpCenterGuideStructure = {
   publisher: {
@@ -103,7 +102,7 @@ export const helpCenterGuideStructure: HelpCenterGuideStructure = {
         element: <AccountSetupGuide />,
         relevantGuides: ["agency-settings", "explore-data"],
       },
-      // Add a new guide within Publisher's directory
+      // Adds a new guide within Publisher's directory
       // Template:
       "new-guide-key": {
         category: "Add Data",
@@ -123,4 +122,4 @@ export const helpCenterGuideStructure: HelpCenterGuideStructure = {
 ---
 
 ### If you'd like to remove a Help Center guide:
-Please delete the relevant guide object from within the `helpCenterGuideStructure` object in `HelpCenterSetup.tsx` and the guide component in `HelpCenterGuides.tsx`.
+Please delete the relevant guide object from within the `helpCenterGuideStructure` object in `HelpCenterSetup.tsx` and the guide component in the `Guides` folder.
